NTBS-3078: Switch to building WebApplication, remove startup.cs
