+++ ntbs-integration-tests/ContactDetailsPages/ContactDetailsPageTests.cs
@@ -11,7 +11,7 @@ namespace ntbs_integration_tests.ContactDetailsPages
 {
     public class ContactDetailsPageTests : TestRunnerBase
     {
-        public ContactDetailsPageTests(NtbsWebApplicationFactory<Startup> factory)
+        public ContactDetailsPageTests(NtbsWebApplicationFactory<Program> factory)
             : base(factory) { }
 
         [Fact]
