+++ ntbs-integration-tests/ContactDetailsPages/EditContactDetailsTests.cs
@@ -13,7 +13,7 @@ namespace ntbs_integration_tests.ContactDetailsPages
 {
     public class EditContactDetailsTests : TestRunnerBase
     {
-        public EditContactDetailsTests(NtbsWebApplicationFactory<Startup> factory)
+        public EditContactDetailsTests(NtbsWebApplicationFactory<Program> factory)
             : base(factory) { }
 
         [Fact]
