+++ ntbs-integration-tests/NotificationPages/DeletePageTests.cs
@@ -14,7 +14,7 @@ namespace ntbs_integration_tests.NotificationPages
     {
         protected override string NotificationSubPath => NotificationSubPaths.Delete;
 
-        public DeletePageTests(NtbsWebApplicationFactory<Startup> factory) : base(factory) { }
+        public DeletePageTests(NtbsWebApplicationFactory<Program> factory) : base(factory) { }
 
         public static IList<Notification> GetSeedingNotifications()
         {
