+++ ntbs-integration-tests/NotificationPages/MBovisOccupationExposurePageTests.cs
@@ -15,7 +15,7 @@ namespace ntbs_integration_tests.NotificationPages
     {
         protected override string NotificationSubPath => NotificationSubPaths.EditMBovisOccupationExposures;
 
-        public MBovisOccupationExposurePageTests(NtbsWebApplicationFactory<Startup> factory) : base(factory)
+        public MBovisOccupationExposurePageTests(NtbsWebApplicationFactory<Program> factory) : base(factory)
         {
         }
