+++ ntbs-integration-tests/NotificationPages/MDRDetailsPageTests.cs
@@ -13,7 +13,7 @@ namespace ntbs_integration_tests.NotificationPages
     {
         protected override string NotificationSubPath => NotificationSubPaths.EditMDRDetails;
 
-        public MdrDetailsPageTests(NtbsWebApplicationFactory<Startup> factory) : base(factory) { }
+        public MdrDetailsPageTests(NtbsWebApplicationFactory<Program> factory) : base(factory) { }
 
         [Fact]
         public async Task Post_ReturnsPageWithFieldsRequiredErrors_IfExposureYesSelected()
