+++ ntbs-integration-tests/NotificationPages/NotificationSummaryTests.cs
@@ -8,7 +8,7 @@ namespace ntbs_integration_tests.NotificationPages
 {
     public class NotificationSummaryTests : TestRunnerBase
     {
-        public NotificationSummaryTests(NtbsWebApplicationFactory<Startup> factory) : base(factory) { }
+        public NotificationSummaryTests(NtbsWebApplicationFactory<Program> factory) : base(factory) { }
 
         private static string PageRoute(string notificationId) => $"/NotificationSummary/{notificationId}";
