+++ ntbs-integration-tests/NotificationPages/TestResultsEditPageTests.cs
@@ -12,7 +12,7 @@ namespace ntbs_integration_tests.NotificationPages
     {
         protected override string NotificationSubPath => NotificationSubPaths.EditTestResults;
 
-        public TestResultsEditPageTest(NtbsWebApplicationFactory<Startup> factory) : base(factory)
+        public TestResultsEditPageTest(NtbsWebApplicationFactory<Program> factory) : base(factory)
         {
         }
