+++ ntbs-integration-tests/ServiceDirectory/ServiceDirectorySearchTest.cs
@@ -9,7 +9,7 @@ namespace ntbs_integration_tests.ServiceDirectory
 {
     public class ServiceDirectorySearchTest : TestRunnerBase
     {
-        public ServiceDirectorySearchTest(NtbsWebApplicationFactory<Startup> factory) : base(factory) { }
+        public ServiceDirectorySearchTest(NtbsWebApplicationFactory<Program> factory) : base(factory) { }
 
         public const string PageRoute = "ServiceDirectory";
