+++ ntbs-integration-tests/TestRunnerNotificationBase.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using ntbs_service;
 using ntbs_service.Helpers;
 
 namespace ntbs_integration_tests
@@ -8,7 +7,7 @@ namespace ntbs_integration_tests
     {
         protected virtual string NotificationSubPath { get; }
 
-        protected TestRunnerNotificationBase(NtbsWebApplicationFactory<Startup> factory) : base(factory) { }
+        protected TestRunnerNotificationBase(NtbsWebApplicationFactory<Program> factory) : base(factory) { }
 
         protected string GetCurrentPathForId(int id)
         {
