+++ ntbs-integration-tests/TransferPages/ActionTransferPageTests.cs
@@ -16,7 +16,7 @@ namespace ntbs_integration_tests.TransferPages
     public class ActionTransferPageTests : TestRunnerNotificationBase
     {
         protected override string NotificationSubPath => NotificationSubPaths.ActionTransferRequest;
-        public ActionTransferPageTests(NtbsWebApplicationFactory<Startup> factory) : base(factory) { }
+        public ActionTransferPageTests(NtbsWebApplicationFactory<Program> factory) : base(factory) { }
 
         public static IList<Notification> GetSeedingNotifications()
         {
